requirements-locked.txt
cbor2==5.2.0
colorzero==1.1
Deprecated==1.2.10
evdev==1.3.0
gpiozero==1.5.1
luma.core==1.17.1
luma.led-matrix==1.5.0
Pillow==7.2.0
pyftdi==0.52.0
pynput==1.7.1
pyserial==3.4
python-xlib==0.28
pyusb==1.1.0
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
simpleaudio==1.0.4
six==1.15.0
smbus2==0.3.0
spidev==3.4
wrapt==1.12.1
ws2812==1.0.0
